dynamic_library: use RTLD_LAZY by default
You can still override it manually, but this seems to be the recommended
default mode. From OS X's `man dlopen`:

"RTLD_LAZY is normally preferred, for reasons of efficiency.  However,
RTLD_NOW is useful to ensure that any undefined symbols are discovered during
the call to dlopen()."

This should at least prevent the "undefined symbol" errors in #80.