function: use a WeakMap to store Callback buffers
See #241.

Before, if you were using the `ffi.Function` "type" and passed
a callback function in such that a C library stored the function
pointer for later, then we want to prevent that callback pointer
from being garbage collected until _after_ the JS callback function
is collected.

However, the the WeakMap is in effect just a regular Map, since
the Callback buffers retain a reference to the JS function, which
is also the `key` we are using for the map. The result is that
in order for the callback pointer memory to be cleaned up you
need to not only nullify the JS callback reference, but *also*
the FunctionType instance references.

This is tricky, since if you pass the FunctionType to
`ffi.ForeignFunction` then the foreign function instance will
also retain a reference to the function type. So you need to
nullify the foreign function reference as well if you are wanting
to free the callback pointer's memory.